ai-services/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.0
pydantic==2.5.0
python-dotenv==1.0.0
pymongo==4.6.0
playwright==1.41.0
beautifulsoup4==4.12.0
requests==2.31.0
spacy==3.7.0
python-multipart==0.0.6
aiohttp==3.9.0